+++ src/main/java/uy/com/equipos/panelmanagement/data/SurveyPanelistParticipation.java
@@ -5,6 +5,11 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
+import java.util.HashSet;
+import java.util.Set;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.OneToMany;
+
 
 @Entity
 public class SurveyPanelistParticipation extends AbstractEntity {
@@ -61,5 +66,16 @@ public class SurveyPanelistParticipation extends AbstractEntity {
 
     public void setCompleted(boolean completed) {
         this.completed = completed;
+    }
+
+    @OneToMany(mappedBy = "surveyPanelistParticipation", cascade = CascadeType.ALL, orphanRemoval = true)
+    private Set<MessageTask> messageTasks = new HashSet<>();
+
+    public Set<MessageTask> getMessageTasks() {
+        return messageTasks;
+    }
+
+    public void setMessageTasks(Set<MessageTask> messageTasks) {
+        this.messageTasks = messageTasks;
     }
 }
